Add compiling functions with args to assembly
Each function call has its own stack frame, to prevent memory address clashes
between functions. I adjust the base pointer and push the arguments onto the
stack before calling the function.

The user has to be sure to match the number of arguments with the number of
parameters. If mismatched it will cause undefined behaviour.